Change access to the KalturaCaptionAsset class.
When creating a caption asset, use the KalturaObjectFactory to grab the
KalturaCaptionAsset class.  This is a huge mind bender because that
class resides in the same plugin module, yet somehow has a different
memory address and will cause downstream use of isinstance() to blow up.

For example, basic use such as the following will blow up with the error
shwon below:

from KalturaPlugins.KalturaCaptionClientPlugin import KalturaCaptionAsset
caption_asset = KalturaCaptionAsset()
[...]
caption_entry = caption_service.add(video_id, caption_asset)

  File "/Users/berto/.virtualenvs/cielo24/src/kaltura/KalturaClientBase.py", line 340, in create
    raise KalturaClientException("Unexpected object type '%s'" % objType, KalturaClientException.ERROR_INVALID_OBJECT_TYPE)
KalturaClientException: Unexpected object type 'KalturaCaptionAsset' (-7)

Further inspection shows that the KalturaCaptionAsset that is being
passed in as a verifier has a different memory address than the one that
is being used to create the object.  Why?  No idea, there's a lot of
non-pythonic business going on in this client library.